+++ packages/carp_audio_package/lib/audio.dart
@@ -17,8 +17,9 @@ import 'package:carp_core/carp_core.dart';
 import 'package:carp_mobile_sensing/carp_mobile_sensing.dart';
 
 part 'audio_probe.dart';
+part 'video_probe.dart';
 part 'noise_probe.dart';
-part 'audio_datum.dart';
+part 'datums.dart';
 part 'audio_measures.dart';
 part 'audio_package.dart';
 part 'audio.g.dart';